vessence/tts_engine/v1/requirements.txt
# TTS Server dependencies
# Install torch+torchaudio with CUDA first:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu128
coqui-tts>=0.24.0
fastapi>=0.100.0
uvicorn>=0.20.0
numpy
